ref: decouple components for other LMS integration
