+++ src/main/java/com/capstone/lms_service/service/CourseService.java
@@ -1,0 +1,7 @@
+package com.capstone.lms_service.service;
+
+import org.common.event.CreateSkillEvent;
+
+public interface CourseService {
+    String createCourse(CreateSkillEvent skillEvent);
+}
