+++ src/main/java/com/capstone/lms_service/service/impl/MoodleCourseService.java
@@ -1,0 +1,13 @@
+package com.capstone.lms_service.service.impl;
+
+import com.capstone.lms_service.service.CourseService;
+import org.common.event.CreateSkillEvent;
+
+public class MoodleCourseService implements CourseService {
+
+    @Override
+    public String createCourse(CreateSkillEvent skillEvent) {
+        //TODO
+        return null;
+    }
+}
